+++ src/interfaces/node.cpp
@@ -25,6 +25,7 @@
 #include <privatesend.h>
 #include <rpc/server.h>
 #include <scheduler.h>
+#include <shutdown.h>
 #include <sync.h>
 #include <txmempool.h>
 #include <ui_interface.h>
