+++ src/privatesend-server.cpp
@@ -7,12 +7,12 @@
 #include <consensus/tx_verify.h>
 #include <consensus/validation.h>
 #include <core_io.h>
-#include <init.h>
 #include <masternode-sync.h>
 #include <masternodeman.h>
 #include <netmessagemaker.h>
 #include <scheduler.h>
 #include <script/interpreter.h>
+#include <shutdown.h>
 #include <txmempool.h>
 #include <util.h>
 #include <utilmoneystr.h>
