+++ src/wallet/wallet.cpp
@@ -13,7 +13,6 @@
 #include <consensus/consensus.h>
 #include <consensus/validation.h>
 #include <fs.h>
-#include <init.h>
 #include <key.h>
 #include <key_io.h>
 #include <keystore.h>
@@ -26,6 +25,7 @@
 #include <primitives/transaction.h>
 #include <reverse_iterator.h>
 #include <script/script.h>
+#include <shutdown.h>
 #include <timedata.h>
 #include <txmempool.h>
 #include <utilmoneystr.h>
